7/3/23

+++ Frontend/src/app/dashboard/components/pages/vendors/vendors.component.ts
@@ -43,11 +43,23 @@ export class VendorsComponent implements OnInit {
   }
 
   DeleteVender(data : number){
-    this.dashboardService.DeleteVender(data).subscribe(
-      (res) => {
-        console.log(JSON.stringify(res));
-        this.vendors = res
-      });
+    
+    if(confirm("Are u sure to delete")){
+      this.dashboardService.DeleteVender(data).subscribe(
+        (res) => {
+          console.log(JSON.stringify(res));
+          this.vendors = res
+        });
+    }
+    
+  }
+
+  EditVendor(data : number){
+    console.log(data + " : "+JSON.stringify(data));
+    this.dashboardService.GetVendor(data).subscribe((res:IVendor) => {
+      this.dashboardService.EmitVendor(res);
+    });
+    this.router.navigate(['/dashboard/AddVendor']);
   }
 
 }
